Telephony: Compilation fix for subId conversion from long to int.
- Fix for compilation issues.

+++ src/com/android/dialer/calllog/CallLogQueryHandler.java
@@ -208,7 +208,7 @@ public class CallLogQueryHandler extends NoNullCursorAsyncQueryHandler {
         }
 
         if (slotId > CALL_SIM_ALL) {
-            long[] subId = SubscriptionManager.getSubId(slotId);
+            int[] subId = SubscriptionManager.getSubId(slotId);
             if (subId != null && subId.length >= 1) {
                 if (where.length() > 0) {
                     where.append(" AND ");
