+++ src/com/android/dialer/calllog/PhoneNumberUtilsWrapper.java
@@ -59,7 +59,7 @@ public class PhoneNumberUtilsWrapper {
      * Returns true if the given number is the number of the configured voicemail of the subId.
      *  To be able to mock-out this, it is not a static method.
      */
-    public boolean isVoicemailNumber(long subId, CharSequence number) {
+    public boolean isVoicemailNumber(int subId, CharSequence number) {
         return number!= null && PhoneNumberUtils.isVoiceMailNumber(subId, number.toString());
     }
